versioning.system.cvss, web.jsf, web.debug, web.core.syntax, properties, websvc.restkit. web.struts and subversion tests now compile

+++ versioning.system.cvss/test/qa-functional/src/org/netbeans/test/cvsmodule/TestKit.java
@@ -43,24 +43,32 @@ package org.netbeans.test.cvsmodule;
 import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileInputStream;
+import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.FileReader;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.util.Arrays;
+import java.util.logging.Handler;
+import java.util.logging.Level;
+import java.util.logging.Logger;
+import java.util.zip.ZipEntry;
+import java.util.zip.ZipInputStream;
 import javax.swing.JCheckBoxMenuItem;
 import org.netbeans.jellytools.EditorOperator;
 import org.netbeans.jellytools.MainWindowOperator;
 import org.netbeans.jellytools.NbDialogOperator;
-import org.netbeans.jellytools.NewFileNameLocationStepOperator;
+import org.netbeans.jellytools.NewJavaFileNameLocationStepOperator;
 import org.netbeans.jellytools.NewFileWizardOperator;
-import org.netbeans.jellytools.NewProjectNameLocationStepOperator;
+import org.netbeans.jellytools.NewJavaProjectNameLocationStepOperator;
 import org.netbeans.jellytools.NewProjectWizardOperator;
 import org.netbeans.jellytools.ProjectsTabOperator;
 import org.netbeans.jellytools.nodes.Node;
+import org.netbeans.jemmy.EventTool;
 import org.netbeans.jemmy.JemmyProperties;
 import org.netbeans.jemmy.QueueTool;
+import org.netbeans.jemmy.TimeoutExpiredException;
 import org.netbeans.jemmy.operators.JCheckBoxMenuItemOperator;
 import org.netbeans.jemmy.operators.JCheckBoxOperator;
 import org.netbeans.jemmy.operators.JMenuBarOperator;
@@ -73,11 +81,14 @@ import org.netbeans.jemmy.operators.JTextFieldOperator;
  * @author Peter Pis
  */
 public class TestKit {
+
     final static String MODIFIED_COLOR = "#0000FF";
     final static String NEW_COLOR = "#008000";
     final static String CONFLICT_COLOR = "#FF0000";
     final static String IGNORED_COLOR = "#999999";
-    
+    public final static String LOGGER_NAME = "org.netbeans.modules.versioning.system.cvss.t9y";
+    public static int TIME_OUT = 15;
+
     public static File createTmpFolder(String prefix) throws IOException {
         String tmpDir = System.getProperty("java.io.tmpdir");  // NOI18N
         File tmpFolder = new File(tmpDir);
@@ -91,17 +102,19 @@ public class TestKit {
         };
         return tmp;
     }
-    
+
     public static void deleteRecursively(File dir) {
-        if (dir.isDirectory()) {
-            String[] files = dir.list();
-            for (int i = 0; i < files.length; i++) {
-                deleteRecursively(new File(dir, files[i]));  // RECURSION
+        if (dir != null) {
+            if (dir.isDirectory()) {
+                String[] files = dir.list();
+                for (int i = 0; i < files.length; i++) {
+                    deleteRecursively(new File(dir, files[i]));  // RECURSION
+                }
             }
+            dir.delete();
         }
-        dir.delete();
     }
-    
+
     public static void pseudoVersion(File file) throws Exception {
         File CVSdir = new File(file, "CVS");
         File Entries = new File(CVSdir, "Entries");
@@ -126,18 +139,22 @@ public class TestKit {
         out.flush();
         out.close();
     }
-    
+
     public static File prepareProject(String category, String project, String project_name, String edit_file) throws Exception {
         //create temporary folder for test
+//        JemmyProperties.setCurrentTimeout("ComponentOperator.WaitComponentTimeout", 36000);
+//        JemmyProperties.setCurrentTimeout("DialogWaiter.WaitDialogTimeout", 36000);
         String folder = "" + System.currentTimeMillis();
         File file = new File("/tmp", folder); // NOI18N
         file.mkdirs();
         //PseudoVersioned project
+        if (System.getProperty("os.name").indexOf("Mac") > -1);
+            NewProjectWizardOperator.invoke().close();
         NewProjectWizardOperator npwo = NewProjectWizardOperator.invoke();
         npwo.selectCategory(category);
         npwo.selectProject(project);
         npwo.next();
-        NewProjectNameLocationStepOperator npnlso = new NewProjectNameLocationStepOperator();
+        NewJavaProjectNameLocationStepOperator npnlso = new NewJavaProjectNameLocationStepOperator();
         new JTextFieldOperator(npnlso, 1).setText(file.getAbsolutePath()); // NOI18N
         new JTextFieldOperator(npnlso, 0).setText(project_name); // NOI18N
         //new JTextFieldOperator(npnlso, 2).setText(folder); // NOI18N
@@ -146,7 +163,7 @@ public class TestKit {
         // wait classpath scanning finished
         return file;
     }
-    
+
     public static void removeAllData(String project_name) {
         Node rootNode = new ProjectsTabOperator().getProjectRootNode(project_name);
         rootNode.performPopupActionNoBlock("Delete Project");
@@ -156,9 +173,9 @@ public class TestKit {
         ndo.yes();
         ndo.getTimeouts().setTimeout("ComponentOperator.WaitStateTimeout", 30000);
         ndo.waitClosed();
-        //TestKit.deleteRecursively(file);
+    //TestKit.deleteRecursively(file);
     }
-    
+
     public static void unversionProject(File file, String project_name) throws Exception {
         File folder_CVS = new File(file, project_name + File.separator + "src" + File.separator + project_name.toLowerCase() + File.separator + "CVS");
         //System.out.println("File: " + file);
@@ -168,16 +185,16 @@ public class TestKit {
         eo.insert("//Comment\n");
         eo.save();
     }
-    
+
     public static void createNewElements(String projectName) {
         String pack = "xx";
-        
+
         NewFileWizardOperator nfwo = NewFileWizardOperator.invoke();
         nfwo.selectProject(projectName);
         nfwo.selectCategory("Java");
         nfwo.selectFileType("Java Package");
         nfwo.next();
-        NewFileNameLocationStepOperator nfnlso = new NewFileNameLocationStepOperator();
+        NewJavaFileNameLocationStepOperator nfnlso = new NewJavaFileNameLocationStepOperator();
         nfnlso.txtObjectName().clearText();
         nfnlso.txtObjectName().typeText(pack);
         nfnlso.finish();
@@ -187,7 +204,7 @@ public class TestKit {
         nfwo.selectCategory("Java");
         nfwo.selectFileType("Java Class");
         nfwo.next();
-        nfnlso = new NewFileNameLocationStepOperator();
+        nfnlso = new NewJavaFileNameLocationStepOperator();
         nfnlso.txtObjectName().clearText();
         nfnlso.txtObjectName().typeText("NewClass");
         nfnlso.selectPackage(pack);
@@ -198,22 +215,22 @@ public class TestKit {
         nfwo.selectCategory("Swing GUI Forms");
         nfwo.selectFileType("JFrame Form");
         nfwo.next();
-        nfnlso = new NewFileNameLocationStepOperator();
+        nfnlso = new NewJavaFileNameLocationStepOperator();
         nfnlso.txtObjectName().clearText();
         nfnlso.txtObjectName().typeText("NewJFrame");
         nfnlso.selectPackage(pack);
         nfnlso.finish();
     }
-    
+
     public static void createNewElementsCommitCvs11(String projectName) {
         String pack = "xx";
-        
+
         NewFileWizardOperator nfwo = NewFileWizardOperator.invoke();
         nfwo.selectProject(projectName);
         nfwo.selectCategory("Java");
         nfwo.selectFileType("Java Package");
         nfwo.next();
-        NewFileNameLocationStepOperator nfnlso = new NewFileNameLocationStepOperator();
+        NewJavaFileNameLocationStepOperator nfnlso = new NewJavaFileNameLocationStepOperator();
         nfnlso.txtObjectName().clearText();
         nfnlso.txtObjectName().typeText(pack);
         nfnlso.finish();
@@ -223,7 +240,7 @@ public class TestKit {
         nfwo.selectCategory("Java");
         nfwo.selectFileType("Java Class");
         nfwo.next();
-        nfnlso = new NewFileNameLocationStepOperator();
+        nfnlso = new NewJavaFileNameLocationStepOperator();
         nfnlso.txtObjectName().clearText();
         nfnlso.txtObjectName().typeText("NewClass");
         nfnlso.selectPackage(pack);
@@ -234,7 +251,7 @@ public class TestKit {
         nfwo.selectCategory("Java");
         nfwo.selectFileType("Java Class");
         nfwo.next();
-        nfnlso = new NewFileNameLocationStepOperator();
+        nfnlso = new NewJavaFileNameLocationStepOperator();
         nfnlso.txtObjectName().clearText();
         nfnlso.txtObjectName().typeText("NewClass2");
         nfnlso.selectPackage(pack);
@@ -245,7 +262,7 @@ public class TestKit {
         nfwo.selectCategory("Java");
         nfwo.selectFileType("Java Class");
         nfwo.next();
-        nfnlso = new NewFileNameLocationStepOperator();
+        nfnlso = new NewJavaFileNameLocationStepOperator();
         nfnlso.txtObjectName().clearText();
         nfnlso.txtObjectName().typeText("NewClass3");
         nfnlso.selectPackage(pack);
@@ -256,33 +273,33 @@ public class TestKit {
         nfwo.selectCategory("Java");
         nfwo.selectFileType("Java Class");
         nfwo.next();
-        nfnlso = new NewFileNameLocationStepOperator();
+        nfnlso = new NewJavaFileNameLocationStepOperator();
         nfnlso.txtObjectName().clearText();
         nfnlso.txtObjectName().typeText("NewClass4");
         nfnlso.selectPackage(pack);
         nfnlso.finish();
     }
-    
+
     public static void createNewElementsCommitCvs12(String projectName) {
         String pack = "aa";
-        
+
         NewFileWizardOperator nfwo = NewFileWizardOperator.invoke();
         nfwo.selectProject(projectName);
         nfwo.selectCategory("Java");
         nfwo.selectFileType("Java Package");
         nfwo.next();
-        NewFileNameLocationStepOperator nfnlso = new NewFileNameLocationStepOperator();
+        NewJavaFileNameLocationStepOperator nfnlso = new NewJavaFileNameLocationStepOperator();
         nfnlso.txtObjectName().clearText();
         nfnlso.txtObjectName().typeText("aa");
         nfnlso.finish();
-        
+
         //
         nfwo = NewFileWizardOperator.invoke();
         nfwo.selectProject(projectName);
         nfwo.selectCategory("Java");
         nfwo.selectFileType("Java Package");
         nfwo.next();
-        nfnlso = new NewFileNameLocationStepOperator();
+        nfnlso = new NewJavaFileNameLocationStepOperator();
         nfnlso.txtObjectName().clearText();
         nfnlso.txtObjectName().typeText("bb");
         nfnlso.finish();
@@ -292,19 +309,19 @@ public class TestKit {
         nfwo.selectCategory("Java");
         nfwo.selectFileType("Java Class");
         nfwo.next();
-        nfnlso = new NewFileNameLocationStepOperator();
+        nfnlso = new NewJavaFileNameLocationStepOperator();
         nfnlso.txtObjectName().clearText();
         nfnlso.txtObjectName().typeText("NewClass");
         nfnlso.selectPackage("aa");
         nfnlso.finish();
-        
+
         //
         nfwo = NewFileWizardOperator.invoke();
         nfwo.selectProject(projectName);
         nfwo.selectCategory("Java");
         nfwo.selectFileType("Java Class");
         nfwo.next();
-        nfnlso = new NewFileNameLocationStepOperator();
+        nfnlso = new NewJavaFileNameLocationStepOperator();
         nfnlso.txtObjectName().clearText();
         nfnlso.txtObjectName().typeText("NewClass2");
         nfnlso.selectPackage("aa");
@@ -315,7 +332,7 @@ public class TestKit {
         nfwo.selectCategory("Java");
         nfwo.selectFileType("Java Class");
         nfwo.next();
-        nfnlso = new NewFileNameLocationStepOperator();
+        nfnlso = new NewJavaFileNameLocationStepOperator();
         nfnlso.txtObjectName().clearText();
         nfnlso.txtObjectName().typeText("NewClass");
         nfnlso.selectPackage("bb");
@@ -326,17 +343,18 @@ public class TestKit {
         nfwo.selectCategory("Java");
         nfwo.selectFileType("Java Class");
         nfwo.next();
-        nfnlso = new NewFileNameLocationStepOperator();
+        nfnlso = new NewJavaFileNameLocationStepOperator();
         nfnlso.txtObjectName().clearText();
         nfnlso.txtObjectName().typeText("NewClass2");
         nfnlso.selectPackage("bb");
         nfnlso.finish();
     }
-    
+
     public static int compareThem(Object[] expected, Object[] actual, boolean sorted) {
         int result = 0;
-        if (expected == null || actual == null)
+        if (expected == null || actual == null) {
             return -1;
+        }
         if (sorted) {
             if (expected.length != actual.length) {
                 return -1;
@@ -364,15 +382,16 @@ public class TestKit {
         }
         return result;
     }
-    
+
     public static String getCVSroot(File cvsFolder) {
         String root = new String();
         if (cvsFolder.isDirectory()) {
             try {
                 BufferedReader br = new BufferedReader(new FileReader(cvsFolder + File.separator + "Root"));
                 root = br.readLine();
-                if (root == null || root.length() == 0)
+                if (root == null || root.length() == 0) {
                     return "";
+                }
                 br.close();
             } catch (IOException e) {
                 return "";
@@ -380,59 +399,148 @@ public class TestKit {
         }
         return root;
     }
-    
+
     public static String getColor(String nodeHtmlDisplayName) {
-        
-        if (nodeHtmlDisplayName == null || nodeHtmlDisplayName.length() < 1)
+
+        if (nodeHtmlDisplayName == null || nodeHtmlDisplayName.length() < 1) {
             return "";
+        }
         int hashPos = nodeHtmlDisplayName.indexOf('#');
         nodeHtmlDisplayName = nodeHtmlDisplayName.substring(hashPos);
         hashPos = nodeHtmlDisplayName.indexOf('"');
         nodeHtmlDisplayName = nodeHtmlDisplayName.substring(0, hashPos);
         return nodeHtmlDisplayName;
     }
-    
+
     public static InputStream getStream(String dir, String protocolName) throws Exception {
         File file = new File(dir, protocolName);
         InputStream in = new FileInputStream(file);
         return in;
     }
-    
+
     public static void closeProject(String projectName) {
         try {
-            Node rootNode = new ProjectsTabOperator().getProjectRootNode(projectName);
-            rootNode.performPopupActionNoBlock("Close");
+            JemmyProperties.setCurrentTimeout("ComponentOperator.WaitComponentTimeout", 2000);
             try {
-                Thread.sleep(2000);
-            } catch (InterruptedException ex) {
-                ex.printStackTrace();
+                Node rootNode = new ProjectsTabOperator().getProjectRootNode(projectName);
+                rootNode.performPopupActionNoBlock("Close");
+                new EventTool().waitNoEvent(2000);
+            } catch (Exception e) {
             }
         } catch (Exception e) {
-            
+        } finally {
+            try {
+                JemmyProperties.setCurrentTimeout("ComponentOperator.WaitComponentTimeout", 10000);
+            } catch (Exception e) {
+            }
         }
+
     }
+
+    public static long changeTimeout(String name, long value) {
+        long timeOut = -1;
+        try {
+            timeOut = JemmyProperties.setCurrentTimeout(name, value);
+        } catch (Throwable e) {
+            e.printStackTrace();
+        }
+        return timeOut;
+    }
+
     public static String getStatus(String nodeHtmlDisplayName) {
-        if (nodeHtmlDisplayName == null || nodeHtmlDisplayName.length() < 1)
+        if (nodeHtmlDisplayName == null || nodeHtmlDisplayName.length() < 1) {
             return "";
+        }
         String status;
         int pos1 = nodeHtmlDisplayName.indexOf('[');
         int pos2 = nodeHtmlDisplayName.indexOf(']');
-        if ((pos1 != -1) && (pos2 != -1))
+        if ((pos1 != -1) && (pos2 != -1)) {
             status = nodeHtmlDisplayName.substring(pos1, pos2 + 1);
-        else
+        } else {
             status = "";
+        }
         return status;
     }
-    
+
     public static void showStatusLabels() {
         JMenuBarOperator mbo = new JMenuBarOperator(MainWindowOperator.getDefault().getJMenuBar());
         JMenuItemOperator mo = mbo.showMenuItem("View|Show Versioning Labels");
         JCheckBoxMenuItemOperator cbmio = new JCheckBoxMenuItemOperator((JCheckBoxMenuItem) mo.getSource());
-        if (!cbmio.getState())
-            cbmio.push();
+        if (!cbmio.getState()) {
+//            cbmio.setState(true);
+            cbmio.doClick();
+        }
     }
-    
+
+    public static void extractProtocol(File path) {
+        ZipInputStream zis = null;
+        OutputStream os = null;
+        String parent = null;
+
+        try {
+            zis = new ZipInputStream(new FileInputStream(new File(path, "protocols.zip")));
+            ZipEntry ze;
+            File dir;
+
+            while ((ze = zis.getNextEntry()) != null) {
+                if (ze.isDirectory()) {
+                    dir = new File(path, ze.getName());
+                    dir.mkdirs();
+                    continue;
+                } else {
+                    os = new FileOutputStream(new File(path, ze.getName()));
+                }
+                byte[] buffer = new byte[1024];
+                int len;
+                while ((len = zis.read(buffer)) > 0) {
+                    os.write(buffer, 0, len);
+                }
+            }
+        } catch (FileNotFoundException ex) {
+            Logger.getLogger(TestKit.class.getName()).log(Level.SEVERE, null, ex);
+        } catch (Exception ex) {
+            Logger.getLogger(TestKit.class.getName()).log(Level.SEVERE, null, ex);
+        } finally {
+            try {
+                zis.close();
+            } catch (IOException ex) {
+                Logger.getLogger(TestKit.class.getName()).log(Level.SEVERE, null, ex);
+            }
+            try {
+                os.close();
+            } catch (IOException ex) {
+                Logger.getLogger(TestKit.class.getName()).log(Level.SEVERE, null, ex);
+            }
+        }
+    }
+
     public static void waitForQueueEmpty() {
         new QueueTool().waitEmpty(1000);
+    }
+
+    public static boolean waitText(MessageHandler handler) {
+        int i = 0;
+
+        while (!handler.isFinished()) {
+            i++;
+            if (i > TIME_OUT) {
+                throw new TimeoutExpiredException("Text [" + handler.message + "] hasn't been found in reasonable time!");
+            }
+            try {
+                Thread.sleep(500);
+            } catch (InterruptedException ex) {
+                ex.printStackTrace();
+            }
+        }
+        return true;
+    }
+
+    public static void removeHandlers(Logger log) {
+        if (log != null) {
+            Handler[] handlers = log.getHandlers();
+            for (int i = 0; i < handlers.length; i++) {
+                log.removeHandler(handlers[i]);
+            }
+        }
     }
 }
